fix(ai): move tool discovery to global assistant + enforce page agent enabledTools
- Apply stub schemas + tool_search to the global assistant route
  (the correct target — handles all 37 tools every request)
- Revert stubs + tool_search from page AI chat route (wrong surface)
- Enforce page.enabledTools server-side in the chat route: when a
  non-empty allowlist is configured, only those tools reach the model.
  This closes a security gap where agents could execute tools outside
  their configured allowlist via the stub/tool_search mechanism.
- Update test mocks accordingly (add to global, remove from chat)
