+++ apps/web/src/app/api/ai/chat/__tests__/mcp-scope.test.ts
@@ -118,7 +118,6 @@ vi.mock('@/lib/ai/core', () => ({
   getUserLMStudioSettings: vi.fn(),
   getUserGLMSettings: vi.fn(),
   pageSpaceTools: {},
-  pageSpaceToolsStubbed: {},
   extractMessageContent: vi.fn().mockReturnValue('test content'),
   extractToolCalls: vi.fn().mockReturnValue([]),
   extractToolResults: vi.fn().mockReturnValue([]),
@@ -192,10 +191,6 @@ vi.mock('@/lib/ai/core/validate-image-parts', () => ({
 
 vi.mock('@/lib/ai/core/model-capabilities', () => ({
   hasVisionCapability: vi.fn().mockReturnValue(true),
-}));
-
-vi.mock('@/lib/ai/tools/tool-search-tool', () => ({
-  createToolSearchTool: vi.fn().mockReturnValue({}),
 }));
 
 import { authenticateRequestWithOptions, checkMCPPageScope } from '@/lib/auth';
