+++ apps/web/src/app/api/ai/chat/__tests__/stream-socket-events.test.ts
@@ -133,7 +133,6 @@ vi.mock('@/lib/ai/core', () => ({
   getUserLMStudioSettings: vi.fn(),
   getUserGLMSettings: vi.fn(),
   pageSpaceTools: {},
-  pageSpaceToolsStubbed: {},
   extractMessageContent: vi.fn().mockReturnValue('test content'),
   extractToolCalls: vi.fn().mockReturnValue([]),
   extractToolResults: vi.fn().mockReturnValue([]),
@@ -224,10 +223,6 @@ vi.mock('@/lib/ai/core/tool-utils', () => ({
 vi.mock('@/lib/ai/tools/finish-tool', () => ({
   finishTool: {},
   FINISH_TOOL_NAME: 'finish',
-}));
-
-vi.mock('@/lib/ai/tools/tool-search-tool', () => ({
-  createToolSearchTool: vi.fn().mockReturnValue({}),
 }));
 
 vi.mock('@/lib/ai/core/stream-pipe-utils', () => ({
